api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=api --output-file=api/requirements.txt
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
chirho @ git+https://github.com/BasisResearch/chirho.git
    # via cities (setup.py)
click==8.1.7
    # via
    #   typer
    #   uvicorn
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dill==0.3.8
    # via cities (setup.py)
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.114.0
    # via cities (setup.py)
fastapi-cli[standard]==0.0.5
    # via fastapi
filelock==3.16.0
    # via torch
fonttools==4.53.1
    # via matplotlib
fsspec==2024.9.0
    # via torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.8
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via
    #   fastapi
    #   torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.7
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.2
    # via cities (setup.py)
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.1
    # via
    #   cities (setup.py)
    #   contourpy
    #   matplotlib
    #   opt-einsum
    #   pandas
    #   pyro-ppl
    #   scikit-learn
    #   scipy
opt-einsum==3.3.0
    # via pyro-ppl
packaging==24.1
    # via
    #   matplotlib
    #   plotly
pandas==2.2.2
    # via cities (setup.py)
pillow==10.4.0
    # via matplotlib
plotly==5.24.0
    # via cities (setup.py)
psycopg2==2.9.9
    # via cities (setup.py)
pydantic==2.9.1
    # via fastapi
pydantic-core==2.23.3
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.1.4
    # via matplotlib
pyro-api==0.1.2
    # via pyro-ppl
pyro-ppl==1.8.6
    # via
    #   chirho
    #   cities (setup.py)
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via uvicorn
rich==13.8.0
    # via typer
scikit-learn==1.5.1
    # via cities (setup.py)
scipy==1.14.1
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.34
    # via cities (setup.py)
starlette==0.38.5
    # via fastapi
sympy==1.13.2
    # via torch
tenacity==9.0.0
    # via plotly
threadpoolctl==3.5.0
    # via scikit-learn
torch @ https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.4.1%2Bcpu.cxx11.abi-cp312-cp312-linux_x86_64.whl
    # via
    #   cities (setup.py)
    #   pyro-ppl
tqdm==4.66.5
    # via pyro-ppl
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typer
tzdata==2024.1
    # via pandas
uvicorn[standard]==0.30.6
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
